Updated the converter and models
