+++ src/parkapi_sources/parkapi_sources.py
@@ -43,9 +43,7 @@ from .converters import (
     ReutlingenPushConverter,
     StuttgartPushConverter,
     UlmPullConverter,
-    VrnParkAndRideMultiguidePullConverter,
     VrnParkAndRidePullConverter,
-    VrnParkAndRideSonahPullConverter,
     VrsBondorfPullConverter,
     VrsKirchheimPullConverter,
     VrsNeustadtPullConverter,
@@ -96,9 +94,7 @@ class ParkAPISources:
         ReutlingenBikePushConverter,
         StuttgartPushConverter,
         UlmPullConverter,
-        VrnParkAndRideMultiguidePullConverter,
         VrnParkAndRidePullConverter,
-        VrnParkAndRideSonahPullConverter,
         VrsBondorfPullConverter,
         VrsKirchheimPullConverter,
         VrsNeustadtPullConverter,
